Homepage - Schedule - Reformat according to editor config
Use 2 spaces as indentation for the yaml file, which is according to the
editorconfig. This also helps identify mistakes

Also removed the empty 'resources' entries as they are not required to
make a valid schedule entry.

Last, give an example for a tutorial lead that has no entry in the
team.yaml