Writeup section about plotting raster images, colorbars and colormaps
Starting the actual plotting tutorial after some fiddling with the ATL14 Iceland dataset. Begin with showing how to plot the actual raster with 3 lines of code. Next, teach a little bit about colorbar customization. Finally, a brief lesson on choosing the right type of Scientific colourmap for the plot! Also added notebook to the tutorial index.